mcp_servers/db_server/requirements.txt
pymongo==4.13.0
python-dotenv==1.1.0
fastmcp==2.5.1
pydantic==2.11.5
pytest==8.4.0
pytest-asyncio==1.0.0
requests==2.32.0